requirements.txt
aiohttp==3.13.3
beautifulsoup4==4.14.3
customtkinter==5.2.2
Pillow==12.1.1
Requests==2.32.5
